requirements.txt
aiofile==3.8.8
aiohttp==3.10.1
coolname==2.2.0
fake_useragent==1.5.1
fastapi==0.115.0
ImageHash==4.3.1
numpy==2.1.2
passlib==1.7.4
pgvector==0.3.2
Pillow==10.4.0
pydantic==2.9.2
PyJWT==2.3.0
pymilvus==2.4.4
python-dotenv==1.0.1
scikit_learn==1.5.2
scipy==1.14.1
SQLAlchemy==2.0.31
torch==2.4.0
transformers==4.43.3